backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
pymysql
cryptography

# Pydantic
pydantic
pydantic[email]
pydantic-settings

# Auth & Security
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]

# Rate Limiting
slowapi

# Utils
python-dotenv
python-multipart

# Testing
pytest